Langchain/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-google-genai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
fastapi>=0.104.0,<1.0.0
langchain_groq
uvicorn[standard]
langserve>=0.0.30
pydantic>=2.0.0,<3.0.0
starlette
sse_starlette